+++ src/components/index.ts
@@ -5,3 +5,4 @@ export { default as PhotoReselectModal } from './PhotoReselectModal';
 export { default as Switch } from './Switch';
 export { default as AiConvertExampleCard } from './AiConvertExampleCard';
 export { default as InputFormItem } from './InputFormItem';
+export { default as PromptSelectModal } from './PromptSelectModal';
