+++ src/pageContainer/AiConversionPage/index.tsx
@@ -1,7 +1,7 @@
 import { useEffect, useState } from 'react';
 
 import { Plus } from '../../assets';
-import { AiConvertExampleCard, StepButton, Switch } from '../../components';
+import { PromptSelectModal, StepButton, Switch } from '../../components';
 import { type CardType, STEP, type Step } from '../../types';
 import type { PromptType } from '../../types/prompt';
 
@@ -14,11 +14,6 @@ interface AiConversionPageProps {
   handleAiConvert: () => void;
 }
 
-interface convertExampleImagesType {
-  title: PromptType;
-  img: string;
-}
-
 const AiConversionPage = ({
   setStep,
   cardType,
@@ -29,14 +24,6 @@ const AiConversionPage = ({
 }: AiConversionPageProps) => {
   const [isAiConvert, setIsAiConvert] = useState<boolean>(false);
   const [isModalOpen, setIsModalOpen] = useState<boolean>(false);
-
-  const convertExampleImages: convertExampleImagesType[] = [
-    { title: '디즈니', img: '/images/디즈니.png' },
-    { title: '레고', img: '/images/레고.png' },
-    { title: '마인크래프트', img: '/images/마인크래프트.png' },
-    { title: '스누피', img: '/images/스누피.png' },
-    { title: '심슨', img: '/images/심슨.png' },
-  ];
 
   useEffect(() => {
     if (isAiConvert) setIsModalOpen(isAiConvert);
@@ -105,61 +92,14 @@ const AiConversionPage = ({
           </div>
         </div>
       </div>
-
       {/* 프롬프트 선택 */}
-      <div
-        className={`absolute inset-0 z-10 h-full w-full rounded-3xl bg-[rgba(0,0,0,0.25)] ${isModalOpen ? 'block' : 'hidden'}`}
-      >
-        <div className="absolute right-0 bottom-0 left-0 flex h-[30.625rem] flex-col gap-9 rounded-3xl bg-white px-12 pt-12 pb-10">
-          <div className="flex h-28 flex-col gap-4">
-            <h1 className="text-4xl leading-none font-extrabold text-[#222]">
-              AI 변환 키워드 선택
-            </h1>
-            <p className="text-xl leading-[150%] font-medium text-[#666]">
-              AI 변환 키워드를 선택해주세요.
-              <br />
-              선택한 키워드로 아래 사진과 같은 스타일로 변환됩니다.
-            </p>
-          </div>
-
-          <div className="flex gap-4">
-            {convertExampleImages.map((x) => (
-              <AiConvertExampleCard
-                key={x.title}
-                title={x.title}
-                img={x.img}
-                selectedPrompt={selectedPrompt}
-                setSelectedPrompt={setSelectedPrompt}
-              />
-            ))}
-          </div>
-
-          <div className="flex justify-end">
-            <div className="flex gap-6">
-              <StepButton
-                variant="back"
-                onClick={() => {
-                  setIsAiConvert(false);
-                  setIsModalOpen(false);
-                }}
-              >
-                AI로 변환하지 않을래요
-              </StepButton>
-              <StepButton
-                variant="next"
-                onClick={() => {
-                  if (selectedPrompt === null) {
-                    setIsAiConvert(false);
-                  }
-                  setIsModalOpen(false);
-                }}
-              >
-                확인
-              </StepButton>
-            </div>
-          </div>
-        </div>
-      </div>
+      <PromptSelectModal
+        isModalOpen={isModalOpen}
+        selectedPrompt={selectedPrompt}
+        setSelectedPrompt={setSelectedPrompt}
+        setIsAiConvert={setIsAiConvert}
+        setIsModalOpen={setIsModalOpen}
+      />
     </div>
   );
 };
